backend/pyproject.toml
[tool.poetry]
name = "shipments-backend"
version = "0.1.0"
description = ""
authors = ["Illia Papou"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
mypy = "^1.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
